Re-implement verbose SQL logging
